# tests/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminLevel": {
        "description": "Level of a place in the administrative hierarchy (glossary: Place).\n\nMembers are declared from the top of the hierarchy down, and that declaration order\n*is* the ordering: ``country`` is the highest level, ``village`` the lowest.\n\n``DIVISION`` is a **proposed** addition to the glossary's hierarchy:\nGilgit-Baltistan groups its districts into divisions, so the level exists between\n``PROVINCE_OR_REGION`` and ``DISTRICT``. The maintainer has not yet confirmed it\n(open question in ``docs/data-dictionary/geography.md``).\n\nImplements: Value Object.",
        "enum": [
          "country",
          "province_or_region",
          "division",
          "district",
          "tehsil",
          "union_council",
          "village"
        ],
        "title": "AdminLevel",
        "type": "string"
      },
...
      "HazardTypeDetail": {
        "additionalProperties": false,
        "description": "One hazard type with everything a reader may need.\n\nImplements: DTO.\n\nAttributes:\n    id: Surrogate identifier (UUIDv7).\n    code: Stable hazard code.\n    parent_code: Code of the broader type, or ``None`` for a root.\n    labels: Display labels per language.\n    description: Longer explanation per language, if written.\n    alignment: Placement in the IRDR 2014 peril classification.\n    attributes_schema: Registry code of the attribute schema, if any.\n    status: ``active`` or ``retired``.\n    retirement: Why it was retired and what replaces it; ``None`` while active.\n    version: Optimistic-concurrency version.\n    created_at: When the type was created, UTC.\n    updated_at: When it last changed, UTC.",
        "properties": {
          "alignment": {
            "$ref": "#/components/schemas/IrdrAlignment"
          },
          "attributes_schema": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_]{1,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes Schema"
          },
          "code": {
            "pattern": "^[a-z][a-z0-9_]{1,63}$",
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalizedText"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "labels": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "parent_code": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_]{1,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Code"
          },
          "retirement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/yakhnama__modules__hazards__domain__value_objects__RetirementReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/HazardTypeStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "parent_code",
          "labels",
          "description",
          "alignment",
          "attributes_schema",
          "status",
          "retirement",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "HazardTypeDetail",
        "type": "object"
      },
      "HazardTypePage": {
        "additionalProperties": false,
        "description": "One page of hazard types, ordered by code.\n\nImplements: API Schema.\n\nAttributes:\n    items: The hazard types on this page.\n    next_cursor: Cursor of the next page, or ``None`` on the last page.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HazardTypeSummary"
            },
            "maxItems": 200,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "HazardTypePage",
        "type": "object"
      },
      "HazardTypeStatus": {
        "description": "Lifecycle status of a hazard type code.\n\nImplements: Value Object.",
        "enum": [
          "active",
          "retired"
        ],
        "title": "HazardTypeStatus",
        "type": "string"
      },
      "HazardTypeSummary": {
        "additionalProperties": false,
        "description": "One hazard type in a listing.\n\nImplements: DTO.\n\nAttributes:\n    code: Stable hazard code.\n    parent_code: Code of the broader type, or ``None`` for a root.\n    labels: Display labels per language.\n    status: ``active`` or ``retired``.\n    attributes_schema: Registry code of the attribute schema, if any.",
        "properties": {
          "attributes_schema": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_]{1,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes Schema"
          },
          "code": {
            "pattern": "^[a-z][a-z0-9_]{1,63}$",
            "title": "Code",
            "type": "string"
          },
          "labels": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "parent_code": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_]{1,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Code"
          },
          "status": {
            "$ref": "#/components/schemas/HazardTypeStatus"
          }
        },
        "required": [
          "code",
          "parent_code",
          "labels",
          "status",
          "attributes_schema"
        ],
        "title": "HazardTypeSummary",
        "type": "object"
      },
...
      "LocalizedText": {
        "additionalProperties": false,
        "description": "One text in several languages and scripts.\n\nThe mapping is exposed read-only so a frozen value cannot be changed through it.\n\nImplements: Value Object.\n\nAttributes:\n    texts: Text per language code, at least one entry.",
        "properties": {
          "texts": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Texts",
            "type": "object"
          }
        },
        "required": [
          "texts"
        ],
        "title": "LocalizedText",
        "type": "object"
      },
...
      "PlaceDetail": {
        "additionalProperties": false,
        "description": "One place with every name and its lifecycle.\n\nImplements: DTO.\n\nAttributes:\n    id: Stable identity (UUIDv7).\n    code: Stable machine code.\n    level: Administrative level.\n    parent_id: The enclosing place, ``None`` for a country.\n    parent_code: Code of the enclosing place, ``None`` for a country.\n    names: Every name, in insertion order.\n    centroid: Representative WGS84 point, if known.\n    status: ``active``, ``merged`` or ``retired``.\n    status_reason: Why it was merged or retired; ``None`` while active.\n    merged_into_id: The replacing place when merged.\n    version: Optimistic-concurrency version.\n    created_at: When the place was first recorded, UTC.\n    updated_at: When it last changed, UTC.",
        "properties": {
          "centroid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coordinates"
              },
              {
                "type": "null"
              }
            ]
          },
          "code": {
            "pattern": "^[a-z0-9][a-z0-9_.-]{1,63}$",
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/AdminLevel"
          },
          "merged_into_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged Into Id"
          },
          "names": {
            "items": {
              "$ref": "#/components/schemas/PlaceName"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Names",
            "type": "array"
          },
          "parent_code": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9_.-]{1,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Code"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "status": {
            "enum": [
              "active",
              "merged",
              "retired"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "level",
          "parent_id",
          "parent_code",
          "names",
          "centroid",
          "status",
          "status_reason",
          "merged_into_id",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "PlaceDetail",
        "type": "object"
      },
      "PlaceFormat": {
        "description": "Representation asked for with ``?format=``; it overrides ``Accept``.\n\nImplements: API Schema.",
        "enum": [
          "json",
          "geojson"
        ],
        "title": "PlaceFormat",
        "type": "string"
      },
...
      "PlacePage": {
        "additionalProperties": false,
        "description": "One page of places matching a search.\n\nImplements: API Schema.\n\nAttributes:\n    items: The places on this page.\n    next_cursor: Cursor of the next page, or ``None`` on the last page.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaceSummary"
            },
            "maxItems": 200,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "PlacePage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "An OpenID Connect access token from the Yakhnama identity provider.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Yakhnama",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/hazard-types": {
      "get": {
        "description": "List hazard types, ordered by code, active ones only unless asked.\n\nArgs:\n    parameters: Validated filters, cursor and limit.\n    response: Used to set the ``Link`` header of the next page.\n    services: Query services bound by the composition root.\n\nReturns:\n    One page of hazard types.",
        "operationId": "list_hazard_types_api_v1_hazard_types_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_retired",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Retired",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardTypePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "List Hazard Types",
        "tags": [
          "hazards"
        ]
      }
    },
    "/api/v1/hazard-types/{code}": {
      "get": {
        "description": "Return one hazard type, active or retired, with its ``ETag``.\n\nArgs:\n    code: The hazard code, for example ``glof``.\n    response: Used to set the ``ETag`` header.\n    services: Query services bound by the composition root.\n\nReturns:\n    The hazard type.\n\nRaises:\n    NotFoundError: If no hazard type has this code.",
        "operationId": "get_hazard_type_api_v1_hazard_types__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "maxLength": 64,
              "pattern": "^[a-z][a-z0-9_]{1,63}$",
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardTypeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Get Hazard Type",
        "tags": [
          "hazards"
        ]
      }
    },
...
    "/api/v1/places": {
      "get": {
        "description": "Search places by name, as JSON or as a GeoJSON ``FeatureCollection``.\n\nArgs:\n    parameters: Validated search text, filters, cursor and limit.\n    response: Used to set ``Link`` and ``Vary`` on the JSON response.\n    services: Query services bound by the composition root.\n    accept: ``application/geo+json`` selects GeoJSON.\n\nReturns:\n    One page of places in the negotiated representation. The GeoJSON form\n    carries the next page only in the ``Link`` header.",
        "operationId": "list_places_api_v1_places_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 8,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AdminLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PlaceFormat"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "accept",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/geo+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacePage"
                }
              }
            },
            "description": "JSON, or GeoJSON when negotiated"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "List Places",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/places/{place_id}": {
      "get": {
        "description": "Return one place, whatever its status, with its ``ETag``.\n\nArgs:\n    place_id: The place id (UUIDv7).\n    parameters: The optional ``format``.\n    response: Used to set ``ETag`` and ``Vary`` on the JSON response.\n    services: Query services bound by the composition root.\n    accept: ``application/geo+json`` selects a GeoJSON ``Feature``.\n\nReturns:\n    The place in the negotiated representation.\n\nRaises:\n    NotFoundError: If no place has this id.",
        "operationId": "get_place_api_v1_places__place_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PlaceFormat"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          },
          {
            "in": "header",
            "name": "accept",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/geo+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDetail"
                }
              }
            },
            "description": "JSON, or GeoJSON when negotiated"
          },
          "401": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          },
          "429": {
            "content": {
              "application/problem+json": {}
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Get Place",
        "tags": [
          "geography"
        ]
      }
    },
...
  }
}